Added styling changes and vuetify config for theme and colors.
